documented internal representation of MoC atoms
